Refatoração inclusão do middleware Menu em um arquivo separado, remoção do middleware somente no arquivo principal app.js

+++ models/product.js
@@ -50,7 +50,7 @@ const product = connection.define('produtos', {
         allowNull: false
     },
     qtd: {
-        type: sequelize.INTEGER,
+        type: sequelize.BIGINT,
         allowNull: false
     },
     propriedadeDivisao: {
